[CWS] add more info to path errors
