+++ pkg/security/resolvers/dentry/resolver.go
@@ -423,7 +423,7 @@ func (dr *Resolver) ResolveFromMap(mountID uint32, inode uint64, pathID uint32, 
 		key.Write(keyBuffer)
 		if err := dr.pathnames.Lookup(keyBuffer, &pathLeaf); err != nil {
 			filenameParts = nil
-			resolutionErr = errDentryPathKeyNotFound
+			resolutionErr = &ErrDentryPathKeyNotFound{PathKey: key}
 			break
 		}
 		depth++
@@ -810,60 +810,6 @@ func (dr *Resolver) Start(manager *manager.Manager) error {
 func (dr *Resolver) Close() error {
 	return fmt.Errorf("couldn't cleanup eRPC memory segment: %w", unix.Munmap(dr.erpcSegment))
 }
-
-// ErrERPCRequestNotProcessed is used to notify that the eRPC request was not processed
-type ErrERPCRequestNotProcessed struct{}
-
-func (err ErrERPCRequestNotProcessed) Error() string {
-	return "erpc_not_processed"
-}
-
-var errERPCRequestNotProcessed ErrERPCRequestNotProcessed
-
-// ErrTruncatedParentsERPC is used to notify that some parents of the path are missing
-type ErrTruncatedParentsERPC struct{}
-
-func (err ErrTruncatedParentsERPC) Error() string {
-	return "truncated_parents_erpc"
-}
-
-var errTruncatedParentsERPC ErrTruncatedParentsERPC
-
-// ErrTruncatedParents is used to notify that some parents of the path are missing
-type ErrTruncatedParents struct{}
-
-func (err ErrTruncatedParents) Error() string {
-	return "truncated_parents"
-}
-
-var errTruncatedParents ErrTruncatedParents
-
-// ErrERPCResolution is used to notify that the eRPC resolution failed
-type ErrERPCResolution struct{}
-
-func (err ErrERPCResolution) Error() string {
-	return "erpc_resolution"
-}
-
-var errERPCResolution ErrERPCResolution
-
-// ErrKernelMapResolution is used to notify that the Kernel maps resolution failed
-type ErrKernelMapResolution struct{}
-
-func (err ErrKernelMapResolution) Error() string {
-	return "map_resolution"
-}
-
-var errKernelMapResolution ErrKernelMapResolution
-
-// ErrDentryPathKeyNotFound is used to notify that the request key is missing from the kernel maps
-type ErrDentryPathKeyNotFound struct{}
-
-func (err ErrDentryPathKeyNotFound) Error() string {
-	return "dentry_path_key_not_found"
-}
-
-var errDentryPathKeyNotFound ErrDentryPathKeyNotFound
 
 // NewResolver returns a new dentry resolver
 func NewResolver(config *config.Config, statsdClient statsd.ClientInterface, e *erpc.ERPC) (*Resolver, error) {
